+++ src/main/java/com/cta4j/bus/prediction/query/StopsPredictionsQuery.java
@@ -57,8 +57,6 @@ public record StopsPredictionsQuery(
     public static Builder builder(List<String> stopIds) {
         Objects.requireNonNull(stopIds);
 
-        stopIds = List.copyOf(stopIds);
-
         return new Builder(stopIds);
     }
 
@@ -92,7 +90,7 @@ public record StopsPredictionsQuery(
         public Builder(List<String> stopIds) {
             Objects.requireNonNull(stopIds);
 
-            this.stopIds = List.copyOf(stopIds);
+            this.stopIds = stopIds;
         }
 
         /**
@@ -105,7 +103,7 @@ public record StopsPredictionsQuery(
         public Builder routeIds(List<String> routeIds) {
             Objects.requireNonNull(routeIds);
 
-            this.routeIds = List.copyOf(routeIds);
+            this.routeIds = routeIds;
 
             return this;
         }