+++ src/main/java/com/cta4j/bus/prediction/query/VehiclesPredictionsQuery.java
@@ -74,7 +74,7 @@ public record VehiclesPredictionsQuery(
         public Builder(List<String> vehicleIds) {
             Objects.requireNonNull(vehicleIds);
 
-            this.vehicleIds = vehicleIds;
+            this.vehicleIds = List.copyOf(vehicleIds);
         }
 
         /**